Renamed the module (and function) 'get_authors_list' to 'get_institution_authors'. Made the necessary changes in other files.
